move BiFunction to ozone common.
