+++ hadoop-ozone/ozone-manager/src/main/java/org/apache/hadoop/ozone/om/request/bucket/acl/OMBucketAddAclRequest.java
@@ -23,7 +23,7 @@ import java.util.List;
 
 import com.google.common.collect.Lists;
 import org.apache.hadoop.ozone.om.OMMetrics;
-import org.apache.hadoop.ozone.om.request.util.BiFunction;
+import org.apache.hadoop.ozone.util.BiFunction;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
